Change file structure of module spec message extraction
