+++ cjwstate/modules/i18n/__init__.py
@@ -1,0 +1,1 @@
+from cjwstate.modules.i18n.catalogs import update_module_catalogs
