+++ cjwstate/modules/i18n/spec_messages.py
@@ -1,75 +1,10 @@
-from cjworkbench.i18n import supported_locales, default_locale
-from cjworkbench.i18n.catalogs.util import (
-    read_po_catalog,
-    write_po_catalog,
-    remove_strings,
-    find_fuzzy_messages,
-    fill_catalog,
-    mark_fuzzy,
-    MessageUID,
-)
-from babel.messages.catalog import Catalog
-from cjwstate.modules.module_loader import ModuleFiles, ModuleSpec
-from typing import FrozenSet, Dict
-import pathlib
 from cjwstate.models.param_spec import ParamSpec, MenuOptionEnum
 from functools import singledispatch
+from typing import Dict
+from cjwstate.modules.module_loader import ModuleSpec
 
 
-def update_module_catalogs(directory: pathlib.Path):
-    module_files = ModuleFiles.load_from_dirpath(directory)  # raise ValueError
-    source_catalog = _build_source_catalog(ModuleSpec.load_from_path(module_files.spec))
-
-    po_path = _po_path(directory, default_locale)
-    try:
-        old_source_catalog = read_po_catalog(po_path)
-    except FileNotFoundError:
-        old_source_catalog = Catalog(default_locale)
-    fuzzy = find_fuzzy_messages(
-        old_catalog=old_source_catalog, new_catalog=source_catalog
-    )
-    write_po_catalog(po_path, source_catalog)
-
-    for locale_id in supported_locales:
-        po_path = _po_path(directory, locale_id)
-        try:
-            old_catalog = read_po_catalog(po_path)
-        except FileNotFoundError:
-            old_catalog = Catalog(locale_id)
-        write_po_catalog(
-            po_path,
-            _merge_nonsource_catalog(locale_id, old_catalog, source_catalog, fuzzy),
-        )
-
-
-def _po_path(basepath: pathlib.Path, locale_id: str) -> pathlib.Path:
-    return basepath / "locale" / locale_id / "messages.po"
-
-
-def _build_source_catalog(spec: ModuleSpec) -> Catalog:
-    messages = _find_spec_messages(spec)
-    # TODO: also find messages in module code
-    source_catalog = Catalog(default_locale)
-    for message_id, source_string in messages.items():
-        if source_string is not None:
-            source_catalog.add(message_id, string=source_string)
-
-    return source_catalog
-
-
-def _merge_nonsource_catalog(
-    locale_id: str,
-    old_catalog: Catalog,
-    source_catalog: Catalog,
-    fuzzy: FrozenSet[MessageUID],
-) -> Catalog:
-    catalog = Catalog(locale_id)
-    fill_catalog(catalog, source_catalog, old_catalog)
-    mark_fuzzy(catalog, fuzzy, old_catalog)
-    return catalog
-
-
-def _find_spec_messages(spec: ModuleSpec) -> Catalog:
+def find_spec_messages(spec: ModuleSpec) -> Dict[str, str]:
     messages = {
         "_spec.name": spec.get("name") or None,
         "_spec.description": spec.get("description") or None,
